Python/requirements.txt
Flask
Flask-PyMongo
bcrypt
PyJWT
python-dotenv
Flask-CORS
requests
mysql-connector-python